feat: model
- Modification des noms de la class et des getters
